# backend/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Voro study management app backend API",
        "title": "Voro API",
        "contact": {
            "name": "Voro Team"
        },
        "version": "1.0"
    },
    "host": "voro-production-405e.up.railway.app",
    "basePath": "/",
    "paths": {
...
        "/api/documents": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quiz"
                ],
                "summary": "List documents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/quiz.DocumentDTO"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quiz"
                ],
                "summary": "Upload a document (PDF)",
                "parameters": [
                    {
                        "type": "file",
                        "description": "PDF file (max 50 MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/quiz.DocumentDTO"
                        }
                    }
                }
            }
        },
        "/api/documents/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "quiz"
                ],
                "summary": "Delete a document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/api/quizzes": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quiz"
                ],
                "summary": "Create a quiz",
                "parameters": [
                    {
                        "description": "Quiz creation options",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/quiz.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/quiz.QuizDTO"
                        }
                    }
                }
            }
        },
        "/api/quizzes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "quiz"
                ],
                "summary": "Get a quiz",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Quiz ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/quiz.QuizDTO"
                        }
                    }
                }
            }
        },
        "/api/setup": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setup"
                ],
                "summary": "Get onboarding state",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/setup.StateDTO"
                        }
                    }
                }
            }
        },
        "/api/setup/steps": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "setup"
                ],
                "summary": "Mark onboarding step complete",
                "parameters": [
                    {
                        "description": "Step to mark (schedule | alarm | exam | notes)",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/setup.StepRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/setup.StateDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "quiz.CreateRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "difficulty": {
                    "type": "string"
                },
                "document_id": {
                    "type": "string"
                },
                "threshold": {
                    "type": "number"
                }
            }
        },
        "quiz.DocumentDTO": {
            "type": "object",
            "properties": {
                "chunk_count": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "source_type": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "quiz.QuestionDTO": {
            "type": "object",
            "properties": {
                "answer_index": {
                    "type": "integer"
                },
                "choices": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "explanation": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "question_text": {
                    "type": "string"
                },
                "source_chunk_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "validation_score": {
                    "type": "number"
                }
            }
        },
        "quiz.QuizDTO": {
            "type": "object",
            "properties": {
                "document_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/quiz.QuestionDTO"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "setup.StateDTO": {
            "type": "object",
            "properties": {
                "alarm": {
                    "type": "boolean"
                },
                "exam": {
                    "type": "boolean"
                },
                "notes": {
                    "type": "boolean"
                },
                "schedule": {
                    "type": "boolean"
                }
            }
        },
        "setup.StepRequest": {
            "type": "object",
            "properties": {
                "step": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}